# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skawld Maintenance API
  version: 0.6.0-phase5
  description: >
    Phase 5 pilot-ready evidence-first maintenance intelligence contract.
    Learned candidates require exact evidence, human review, deterministic
    evaluation, approval authority, and explicit publication. AI quality
    labels are append-oriented. This API exposes no industrial control action
    or permit authority.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/assets:
    get:
      summary: List authorized assets
      tags: [Assets]
      operationId: listAssets
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: query, name: site_id, schema: {type: string, format: uuid}}
        - {in: query, name: q, schema: {type: string}}
        - {in: query, name: cursor, required: false, schema: {type: string}}
        - {in: query, name: page_size, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
      responses:
        "200":
          description: Assets in the principal's authorized scope.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor, has_more]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Asset"}}
                  next_cursor: {type: [string, "null"]}
                  has_more: {type: boolean}
        "401": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
    post:
      summary: Create an asset
      tags: [Assets]
      operationId: createAsset
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateAsset"}
      responses:
        "201":
          description: Asset created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Asset"}
        "200": {$ref: "#/components/responses/IdempotentReplay"}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/assets/{assetID}:
    get:
      summary: Get asset detail
      tags: [Assets]
      operationId: getAsset
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/AssetID"}]
      responses:
        "200":
          description: Authorized asset detail.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Asset"}
        "404": {$ref: "#/components/responses/Problem"}
  /api/v1/assets/{assetID}/criticality-approvals:
    post:
      summary: Approve asset criticality
      tags: [Assets]
      operationId: approveAssetCriticality
      description: Requires both RBAC permission and a matching ApprovalAuthority.
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/AssetID"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CriticalityCommand"}
      responses:
        "201": {description: Human-approved criticality recorded.}
        "403": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
...
  /api/v1/attachments:
    get:
      summary: List attachments for an entity
      tags: [Attachments]
      operationId: listAttachments
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: query, name: entity_kind, required: true, schema: {type: string, enum: [EXECUTION, OBSERVATION, INCIDENT]}}
        - {in: query, name: entity_id, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Attachments with signed download URLs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Attachment"}}
    post:
      summary: Create an attachment manifest
      tags: [Attachments]
      operationId: createAttachmentManifest
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AttachmentManifestCommand"}
      responses:
        "201":
          description: Pending manifest and constrained signed upload request.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Attachment"}
        "409": {$ref: "#/components/responses/Problem"}
  /api/v1/attachments/{attachmentID}/complete:
    post:
      summary: Verify a completed attachment upload
      tags: [Attachments]
      operationId: completeAttachmentUpload
      description: Streams and verifies object size, SHA-256, and magic-byte MIME before availability.
      security: [{sessionCookie: []}, {bearerAuth: []}]
      parameters:
        - {in: path, name: attachmentID, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "200":
          description: Attachment verified and available.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Attachment"}
        "409": {$ref: "#/components/responses/Problem"}
...
  /api/v1/admin/field-definitions:
    get:
      summary: List field definitions for an entity
      operationId: listFieldDefinitions
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - {in: query, name: entity_type, required: false, schema: {type: string, default: incident}}
      responses:
        "200":
          description: Field definitions for the organization and entity.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/FieldDefinition"}
    post:
      summary: Create a field definition
      operationId: createFieldDefinition
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_type, key, label, field_type]
              properties:
                entity_type: {type: string}
                key: {type: string, pattern: '^[a-z][a-z0-9_]{0,63}$'}
                label: {type: string}
                description: {type: string}
                field_type: {type: string, enum: [TEXT, NUMBER, DATE, SELECT, MULTI_SELECT]}
                config: {$ref: "#/components/schemas/FieldDefinition/properties/config"}
                sort_order: {type: integer}
      responses:
        "201":
          description: Created field definition.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FieldDefinition"}
  /api/v1/admin/field-definitions/{fieldID}:
    get:
      summary: Get a field definition
      operationId: getFieldDefinition
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - {in: path, name: fieldID, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: The field definition.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FieldDefinition"}
    patch:
      summary: Update a field definition
      operationId: updateFieldDefinition
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - {in: path, name: fieldID, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, field_type, config, sort_order, expected_version]
              properties:
                label: {type: string}
                description: {type: string}
                field_type: {type: string, enum: [TEXT, NUMBER, DATE, SELECT, MULTI_SELECT]}
                config: {$ref: "#/components/schemas/FieldDefinition/properties/config"}
                sort_order: {type: integer}
                expected_version: {type: integer, format: int64}
      responses:
        "200":
          description: The updated field definition.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FieldDefinition"}
  /api/v1/admin/field-definitions/{fieldID}/retire:
    post:
      summary: Retire a field definition
      operationId: retireFieldDefinition
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - {in: path, name: fieldID, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: The retired field definition.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FieldDefinition"}
  /api/v1/admin/field-definitions/{fieldID}/history:
    get:
      summary: List value history for a field definition
      operationId: fieldDefinitionHistory
      tags: [CustomFields]
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - {in: path, name: fieldID, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Value change history for the field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/HistoryEntry"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: skawld_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token for native/mobile Authorization Code + PKCE.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
    AssetID:
      in: path
      name: assetID
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    IdempotentReplay:
      description: Original successful response replayed for the same command.
      headers:
        Idempotent-Replayed:
          schema: {type: string, const: "true"}
  schemas:
...
    CreateAsset:
      type: object
      additionalProperties: false
      required: [site_id, tag, name, class, source_of_truth]
      properties:
        site_id: {type: string, format: uuid}
        parent_asset_id: {type: string, format: uuid}
        tag: {type: string, minLength: 1}
        name: {type: string, minLength: 1}
        class: {type: string, minLength: 1}
        manufacturer: {type: string}
        model: {type: string}
        source_of_truth:
          type: string
          enum: [OWNED_BY_SKAWLD, EXTERNAL_REFERENCE]
        external_reference:
          type: object
          required: [system, id]
          properties:
            system: {type: string}
            id: {type: string}
            version: {type: string}
        components:
          type: array
          items:
            type: object
            required: [code, name, type]
            properties:
              code: {type: string}
              name: {type: string}
              type: {type: string}
    Asset:
      type: object
      required: [id, organization_id, site_id, tag, name, class, status, source_of_truth, version, components, created_at]
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        site_id: {type: string, format: uuid}
        parent_asset_id: {type: string, format: uuid}
        tag: {type: string}
        name: {type: string}
        class: {type: string}
        manufacturer: {type: string}
        model: {type: string}
        status: {type: string, enum: [ACTIVE, INACTIVE, RETIRED]}
        source_of_truth:
          type: string
          enum: [OWNED_BY_SKAWLD, EXTERNAL_REFERENCE]
        version: {type: integer, format: int64}
        created_at: {type: string, format: date-time}
        components: {type: array, items: {type: object}}
        criticality: {$ref: "#/components/schemas/Criticality"}
    Criticality:
      type: object
      properties:
        rating: {type: string, enum: [A, B, C]}
        safety_impact: {type: integer, minimum: 0, maximum: 5}
        production_impact: {type: integer, minimum: 0, maximum: 5}
        environmental_impact: {type: integer, minimum: 0, maximum: 5}
        financial_impact: {type: integer, minimum: 0, maximum: 5}
        redundancy: {type: string}
        rationale: {type: string}
        approved_by: {type: string, format: uuid}
        approved_at: {type: string, format: date-time}
    CriticalityCommand:
      type: object
      additionalProperties: false
      required: [rating, safety_impact, production_impact, environmental_impact, financial_impact, redundancy, rationale]
      properties:
        rating: {type: string, enum: [A, B, C]}
        safety_impact: {type: integer, minimum: 0, maximum: 5}
        production_impact: {type: integer, minimum: 0, maximum: 5}
        environmental_impact: {type: integer, minimum: 0, maximum: 5}
        financial_impact: {type: integer, minimum: 0, maximum: 5}
        redundancy: {type: string}
        rationale: {type: string, minLength: 1}
...
    FieldDefinition:
      type: object
      required: [id, entity_type, key, label, field_type, status, sort_order, version, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        entity_type: {type: string}
        key: {type: string, pattern: '^[a-z][a-z0-9_]{0,63}$'}
        label: {type: string}
        description: {type: string}
        field_type: {type: string, enum: [TEXT, NUMBER, DATE, SELECT, MULTI_SELECT]}
        config:
          type: object
          properties:
            required: {type: boolean}
            max_length: {type: integer}
            regex: {type: string}
            min: {type: number}
            max: {type: number}
            options:
              type: array
              items:
                type: object
                properties:
                  label: {type: string}
                  value: {type: string}
        status: {type: string, enum: [ACTIVE, RETIRED]}
        sort_order: {type: integer}
        version: {type: integer, format: int64}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        retired_at: {type: [string, "null"], format: date-time}
        has_values: {type: boolean}
        incident_count: {type: integer}
    HistoryEntry:
      type: object
      properties:
        incident_id: {type: string, format: uuid}
        principal_id: {type: string, format: uuid}
        value_before: {}
        value_after: {}
        changed_at: {type: string, format: date-time}
...
    AttachmentManifestCommand:
      type: object
      additionalProperties: false
      required: [site_id, entity_kind, entity_id, client_event_id, original_filename, declared_mime, size_bytes, checksum_sha256]
      properties:
        site_id: {type: string, format: uuid}
        entity_kind: {type: string, enum: [EXECUTION, OBSERVATION, INCIDENT, DOCUMENT_REVISION]}
        entity_id: {type: string, format: uuid}
        client_event_id: {type: string, format: uuid}
        original_filename: {type: string}
        declared_mime:
          type: string
          enum: [image/jpeg, image/png, image/webp, audio/m4a, audio/mp4, audio/mpeg, audio/wav, video/mp4, video/webm, video/quicktime, application/pdf, text/plain]
        size_bytes: {type: integer, minimum: 1, maximum: 52428800}
        checksum_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
    Attachment:
      type: object
      properties:
        id: {type: string, format: uuid}
        organization_id: {type: string, format: uuid}
        site_id: {type: string, format: uuid}
        entity_kind: {type: string}
        entity_id: {type: string, format: uuid}
        client_event_id: {type: string}
        original_filename: {type: string}
        declared_mime: {type: string}
        verified_mime: {type: string}
        size_bytes: {type: integer}
        checksum_sha256: {type: string}
        state:
          type: string
          enum: [PENDING_UPLOAD, UPLOADED, QUARANTINED, AVAILABLE, REJECTED]
        upload_url: {type: string, format: uri}
        upload_headers:
          type: object
          additionalProperties: {type: string}
        upload_expires_at: {type: string, format: date-time}
        download_url: {type: string, format: uri}
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string